requirements.txt
numpy
Pillow
streamlit
streamlit_drawable_canvasz
torch
